Module google_resourceviews1_beta2::api

Structs

Label
The Label to be applied to the resource views.

ListResourceResponseItem
The list response item that contains the resource and end points information.

Operation
An operation resource, used to manage asynchronous API requests.

OperationError
[Output Only] If errors occurred during processing of this operation, this field will be populated.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

OperationList
There is no detailed description.

OperationWarnings
[Output Only] If there are issues with this operation, a warning is returned.

OperationWarningsData
[Output only] Metadata for this warning in key:value format.

ResourceView
The resource view object.

Resourceviews
Central instance to access all Resourceviews related resource activities

ServiceEndpoint
The service endpoint that may be started in a VM.

ZoneOperationGetCall
Retrieves the specified zone-specific operation resource.

ZoneOperationListCall
Retrieves the list of operation resources contained within the specified zone.

ZoneOperationMethods
A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the Resourceviews hub.

ZoneViewAddResourceCall
Add resources to the view.

ZoneViewDeleteCall
Delete a resource view.

ZoneViewGetCall
Get the information of a zonal resource view.

ZoneViewGetServiceCall
Get the service information of a resource view or a resource.

ZoneViewInsertCall
Create a resource view.

ZoneViewListCall
List resource views.

ZoneViewListResourceCall
List the resources of the resource view.

ZoneViewMethods
A builder providing access to all methods supported on zoneView resources. It is not used directly, but through the Resourceviews hub.

ZoneViewRemoveResourceCall
Remove resources from the view.

ZoneViewSetServiceCall
Update the service information of a resource view or a resource.

ZoneViewsAddResourcesRequest
The request to add resources to the resource view.

ZoneViewsGetServiceResponse
There is no detailed description.

ZoneViewsList
The response to a list request.

ZoneViewsListResourcesResponse
The response to a list resource request.

ZoneViewsRemoveResourcesRequest
The request to remove resources from the resource view.

ZoneViewsSetServiceRequest
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.