Struct google_resourceviews1_beta2::ZoneViewMethods

pub struct ZoneViewMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on zoneView resources. It is not used directly, but through the Resourceviews hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_resourceviews1_beta2 as resourceviews1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use resourceviews1_beta2::Resourceviews;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Resourceviews::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_resources(...)`, `delete(...)`, `get(...)`, `get_service(...)`, `insert(...)`, `list(...)`, `list_resources(...)`, `remove_resources(...)` and `set_service(...)`
// to build up your call.
let rb = hub.zone_views();

Methods

impl<'a, C, A> ZoneViewMethods<'a, C, A>

fn remove_resources(&self, request: ZoneViewsRemoveResourcesRequest, project: &str, zone: &str, resource_view: &str) -> ZoneViewRemoveResourceCall<'a, C, A>

Create a builder to help you perform the following task:

Remove resources from the view.

Arguments

request - No description provided.

project - The project name of the resource view.

zone - The zone name of the resource view.

resourceView - The name of the resource view.

fn add_resources(&self, request: ZoneViewsAddResourcesRequest, project: &str, zone: &str, resource_view: &str) -> ZoneViewAddResourceCall<'a, C, A>

Create a builder to help you perform the following task:

Add resources to the view.

Arguments

request - No description provided.

project - The project name of the resource view.

zone - The zone name of the resource view.

resourceView - The name of the resource view.

fn list_resources(&self, project: &str, zone: &str, resource_view: &str) -> ZoneViewListResourceCall<'a, C, A>

Create a builder to help you perform the following task:

List the resources of the resource view.

Arguments

project - The project name of the resource view.

zone - The zone name of the resource view.

resourceView - The name of the resource view.

fn get(&self, project: &str, zone: &str, resource_view: &str) -> ZoneViewGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get the information of a zonal resource view.

Arguments

project - The project name of the resource view.

zone - The zone name of the resource view.

resourceView - The name of the resource view.

fn list(&self, project: &str, zone: &str) -> ZoneViewListCall<'a, C, A>

Create a builder to help you perform the following task:

List resource views.

Arguments

project - The project name of the resource view.

zone - The zone name of the resource view.

fn insert(&self, request: ResourceView, project: &str, zone: &str) -> ZoneViewInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Create a resource view.

Arguments

request - No description provided.

project - The project name of the resource view.

zone - The zone name of the resource view.

fn delete(&self, project: &str, zone: &str, resource_view: &str) -> ZoneViewDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a resource view.

Arguments

project - The project name of the resource view.

zone - The zone name of the resource view.

resourceView - The name of the resource view.

fn set_service(&self, request: ZoneViewsSetServiceRequest, project: &str, zone: &str, resource_view: &str) -> ZoneViewSetServiceCall<'a, C, A>

Create a builder to help you perform the following task:

Update the service information of a resource view or a resource.

Arguments

request - No description provided.

project - The project name of the resource view.

zone - The zone name of the resource view.

resourceView - The name of the resource view.

fn get_service(&self, project: &str, zone: &str, resource_view: &str) -> ZoneViewGetServiceCall<'a, C, A>

Create a builder to help you perform the following task:

Get the service information of a resource view or a resource.

Arguments

project - The project name of the resource view.

zone - The zone name of the resource view.

resourceView - The name of the resource view.