Struct google_resourceviews1_beta2::Operation

pub struct Operation {
    pub status: Option<String>,
    pub insert_time: Option<String>,
    pub warnings: Option<Vec<OperationWarnings>>,
    pub error: Option<OperationError>,
    pub target_id: Option<String>,
    pub target_link: Option<String>,
    pub start_time: Option<String>,
    pub client_operation_id: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub zone: Option<String>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub operation_type: Option<String>,
    pub http_error_message: Option<String>,
    pub progress: Option<i32>,
    pub end_time: Option<String>,
    pub http_error_status_code: Option<i32>,
    pub status_message: Option<String>,
    pub user: Option<String>,
}

An operation resource, used to manage asynchronous API requests.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

remove resources zone views (response)

add resources zone views (response)

delete zone views (response)

get zone operations (response)

insert zone views (response)

set service zone views (response)

Fields

status

[Output Only] Status of the operation.

insert_time

[Output Only] The time that this operation was requested, in RFC3339 text format.

warnings

[Output Only] If there are issues with this operation, a warning is returned.

error

[Output Only] If errors occurred during processing of this operation, this field will be populated.

target_id

[Output Only] Unique target ID which identifies a particular incarnation of the target.

target_link

[Output only] URL of the resource the operation is mutating.

start_time

[Output Only] The time that this operation was started by the server, in RFC3339 text format.

client_operation_id

[Output only] An optional identifier specified by the client when the mutation was initiated. Must be unique for all operation resources in the project.

creation_timestamp

[Output Only] The time that this operation was requested, in RFC3339 text format.

id

[Output Only] Unique identifier for the resource, generated by the server.

kind

[Output only] Type of the resource.

name

[Output Only] Name of the resource.

zone

[Output Only] URL of the zone where the operation resides. Only available when performing per-zone operations.

region

[Output Only] URL of the region where the operation resides. Only available when performing regional operations.

self_link

[Output Only] Server-defined fully-qualified URL for this resource.

operation_type

[Output only] Type of the operation. Operations include insert, update, and delete.

http_error_message

[Output only] If operation fails, the HTTP error message returned.

progress

[Output only] An optional progress indicator that ranges from 0 to 100. There is no requirement that this be linear or support any granularity of operations. This should not be used to guess at when the operation will be complete. This number should be monotonically increasing as the operation progresses.

end_time

[Output Only] The time that this operation was completed, in RFC3339 text format.

http_error_status_code

[Output only] If operation fails, the HTTP error status code returned.

status_message

[Output Only] An optional textual description of the current status of the operation.

user

[Output Only] User who requested the operation, for example: user@example.com.